Add unary-specific precedence and associativity
Associativity still needs to be fleshed out for unary values.
One idea is to have postfix (left-associative) and prefix (right) operators

e.g: ! 4, and 4 !.

	4 ! + 1

would push 4 to the output stack, then ! to the operator stack, then +
to the operator stack. If + is left-associative, then ! can be pushed off
if it has higher precedence. How this affects, however, ! is a binary operator,
so + will be seen as unary instead, and thus, there should be no need for
associativity on unary operators.

The simplest thing to do is to disallow associativity on operators and state
that they must all be prefix operators.